pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plumbline-ai"
version = "0.1.0"
description = "Write-time verification for AI-generated code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Evolving Intelligence AI", email = "nicholas.smith@evolvingintelligence.ai" },
]
keywords = [
    "ai",
    "verification",
    "mcp",
    "code-quality",
    "claude",
    "llm",
    "developer-tools",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "mcp>=1.0,<2.0",
    "typer>=0.12.0,<1.0",
    "rich>=13.0.0,<15.0",
    "pydantic>=2.10,<3.0",
    "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
# Core sidecar engines
context = ["git-sentinel>=0.4.0"]
verify = ["seraph-ai>=0.1.1"]
gate = ["morpheus-mcp>=0.3.0"]

# All engines
all = ["plumbline-ai[context,verify,gate]"]

# Development
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
plumbline = "plumbline.cli.app:app"
plumbline-mcp = "plumbline.mcp.server:main"

[project.urls]
Homepage = "https://plumblinehq.com"
Repository = "https://github.com/evolvingintelligenceai/plumbline"
Issues = "https://github.com/evolvingintelligenceai/plumbline/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/plumbline"]

[tool.hatch.build.targets.sdist]
include = [
    "src/plumbline",
    "tests",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]
exclude = [
    "/.plan_file",
    "/plans",
    "/.env",
    "/.venv",
    "/dist",
    "/build",
    "/.pytest_cache",
    "/.ruff_cache",
    "/.mypy_cache",
    "/.niobe",
    "/.sentinel",
    "**/__pycache__",
    "**/.DS_Store",
]

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers -q"

[tool.mypy]
python_version = "3.10"
strict = true
